Rename file to extended_property_conversion.py
