+++ packages/ni.protobuf.types/src/ni/protobuf/types/scalar_conversion.py
@@ -8,7 +8,7 @@ from nitypes.scalar import Scalar
 from typing_extensions import TypeAlias
 
 import ni.protobuf.types.scalar_pb2 as scalar_pb2
-from ni.protobuf.types.extended_properties_conversion import (
+from ni.protobuf.types.extended_property_conversion import (
     extended_properties_from_protobuf,
     extended_properties_to_protobuf,
 )
